build: update version, change ver format
Removed the 'build' number system from the versions, instead they will use snapshot or release candidate, releases will just be the version itself.
I think this will work with the update checker used on 2.1.7 (at least) since it just blankly checks if the version string is different before the first space in the string occurs (which can be nonexistent).